wallet: extract ListAccounts shared workflow
